backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
argon2-cffi>=23.1.0
python-dotenv>=1.0.0
